Add files via upload
This is the main class that needs to be added to the main branch. it has a method called find which takes two arguments, the first argument is the nth that the user is trying to find its value in the Fibonacci sequence, and the second argument is the array list that contains the values of the Fibonacci sequence. It also has the Main method which generates the Fibonacci sequence which is computed and not hard coded. The main method then asks the user to input the number to search for and finally calls the find method and displays the result.